requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
pydantic==2.6.1
pydantic-core==2.16.2
torch>=2.1.0
torchvision>=0.16.0
